+++ CHANGES.md
@@ -1,6 +1,39 @@
 CHANGES
 =======
 
+Update to follow the following normative changes of the SD-JWT specification.
+
+1. A tilde (`~`) is appended to "Combined Format for Issuance", resulting in
+   that there is no formal difference between (a) Combined Format for Issuance
+   and (b) Combined Format for Presentation without a binding JWT.
+2. The format of the base JSON of Disclosure for an array element has changed
+   from `[salt, [name, index], value]` to `[salt, value]`.
+3. Disclosures for array elements are not put in the `_sd` array. Instead,
+   each array element is replaced with `{"...": "<digest>"}`.
+4. Terminology changes. The terms "Combined Format for Issuance" and "Combined
+   Format for Presentation" are obsoleted, and the entire combined format is
+   called "SD-JWT".
+
+- `Disclosure` class
+  - Added `Disclosure(Object)` constructor.
+  - Added `toArrayElement()` method.
+  - Added `toArrayElement(String)` method.
+  - Removed `Disclosure(String, int, Object)` constructor.
+  - Removed `Disclosure(String, String, int, Object)` constructor.
+  - Removed `getClaimIndex()` method.
+
+- `SDObjectBuilder` class
+  - Removed `putSDClaim(String, int, Object)` method.
+  - Removed `putSDClaim(String, String, int, Object)` method.
+
+- New types
+  - `SDJWT` class
+
+- Removed types
+  - `SDCombinedFormat` class
+  - `SDIssuance` class
+  - `SDPresentation` class
+
 1.1 (2023-05-10)
 ----------------